Updated wg group dependency
