+++ src/edge_node/mod.rs
@@ -1,14 +1,14 @@
 use std::collections::HashMap;
 
 use crossbeam::channel::{Receiver, Sender};
-use wg_2024::{controller::Command, network::NodeId, packet::Packet};
+use wg_2024::{controller::DroneCommand, network::NodeId, packet::Packet};
 
 use crate::fragmentation::Fragmenter;
 
 pub struct EdgeNodeOptions {
     pub id: NodeId,
-    pub sim_contr_send: Sender<Command>,
-    pub sim_contr_recv: Receiver<Command>,
+    pub sim_contr_send: Sender<DroneCommand>,
+    pub sim_contr_recv: Receiver<DroneCommand>,
     pub packet_recv: Receiver<Packet>,
     pub packet_send: HashMap<NodeId, Sender<Packet>>,
 }
