+++ src/tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod send_msg;
+pub mod send_query;
